requirements.txt
fastapi
uvicorn[standard]
azure-search-documents
azure-core
openai
python-dotenv
sqlalchemy

